+++ test/unit/fuzz-tools.test.js
@@ -516,6 +516,7 @@ import * as neosV538 from '../../lib/neos-v538.js';
 import * as islLymphedemaV539 from '../../lib/isl-lymphedema-v539.js';
 import * as ishltRejectionV540 from '../../lib/ishlt-rejection-v540.js';
 import * as rachs1V541 from '../../lib/rachs1-v541.js';
+import * as twstrsSeverityV542 from '../../lib/twstrs-severity-v542.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -977,6 +978,7 @@ const MODULES = {
   'isl-lymphedema-v539.js': islLymphedemaV539,
   'ishlt-rejection-v540.js': ishltRejectionV540,
   'rachs1-v541.js': rachs1V541,
+  'twstrs-severity-v542.js': twstrsSeverityV542,
   'tb-testing.js': tbTesting,
 };